+++ docs/source/changes.md
@@ -24,6 +24,7 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
   {mod}`_pytask.mark_utils`. (Closes {issue}`220`.)
 - {pull}`236` refactors {mod}`_pytask.collect` and places shared functions in
   {mod}`_pytask.collect_utils`.
+- {pull}`238` allows any order of decorators with a `@pytask.mark.task` decorator.
 - {pull}`241` allows to parametrize over single dicts.
 - {pull}`242` removes tasks from global {obj}`_pytask.task_utils.COLLECTED_TASKS` to
   avoid re-collection when the programmatic interface is used.